% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/init.R
\name{EigenNbThreads}
\alias{EigenNbThreads}
\alias{EigenSetNbThreads}
\alias{RcppEigen_throttle_cores}
\alias{RcppEigen_reset_cores}
\title{Throttle (or Reset) (Rcpp)Eigen Core Usage}
\usage{
EigenNbThreads()

EigenSetNbThreads(n)

RcppEigen_throttle_cores(n)

RcppEigen_reset_cores()
}
\arguments{
\item{n}{Integer value of desired cores, default is the value set at package
startup reflecting the smallest value among the total number of available
cores (or one if compiled without OpenMP support), the value of option
\code{Ncpus} and the value of environment variable \code{OMP_THREAD_LIMIT}.}
}
\value{
Only \code{EigenNbThreads()} returns a value, the current value of
the number of cores used. The other functions are invoked for their side
effect of affecting the count of cores used.
}
\description{
Helper functions to throttle use of cores by RcppEigen-internal code.
On package load, the initial value is saved and used to reset the value.
}
\seealso{
\code{\link{RcppEigen-package}}
}
