\name{RcppEigen-package}
\alias{RcppEigen-package}
\alias{RcppEigen}
\docType{package}
\title{
Rcpp/Eigen bridge
}
\description{
  The package eases the use of the Eigen C++ template library for linear
  algebra with Rcpp
}
\details{
  This package contains the header files for the Eigen C++ template
  library.  The typical usage is to install this package and list it in
  the \env{LinkingTo: } line in the \file{DESCRIPTION} file of
  other packages.  The C++ source code and the R source code in this
  package are for illustration only.

  As described at the Eigen project's home page,
  \url{http://eigen.tuxfamily.org}, Eigen is a versatile, fast, reliable
  and elegant collection of C++ classes for linear algebra.
}
\references{\url{http://eigen.tuxfamily.org}}
\keyword{ package }
