\name{RcppEigen-package}
\alias{RcppEigen-package}
\alias{RcppEigen}
\docType{package}
\title{
Rcpp/Eigen bridge
}
\description{
  The package eases the use of the Eigen C++ template library for linear
  algebra with Rcpp
}
\details{
  \tabular{ll}{
    Package: \tab RcppEigen\cr
    Type: \tab Package\cr
    Version: \tab 0.2\cr
    Date: \tab 2011-06-13\cr
    License: \tab GPL(>= 2)\cr
    LazyLoad: \tab yes\cr
  }
}
\references{\url{http://eigen.tuxfamily.org}}
\keyword{ package }
