% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/init.R
\name{EigenNbThreads}
\alias{EigenNbThreads}
\alias{EigenSetNbThreads}
\alias{RcppEigen_throttle_cores}
\alias{RcppEigen_reset_cores}
\title{Throttle (or Reset) (Rcpp)Eigen to Two Cores}
\usage{
EigenNbThreads()

EigenSetNbThreads(n)

RcppEigen_throttle_cores(n)

RcppEigen_reset_cores()
}
\arguments{
\item{n}{Integer value of desired cores, default is two}
}
\value{
Only \code{EigenNbThreads()} returns a value, the current value of
the number of cores used. The other functions are invoked for their side-effect
of affecting the core count.
}
\description{
Helper functions to throttle use of cores by RcppEigen-internal code.
On package load, the initial value is saved and used to reset the value.
}
\seealso{
\code{\link{RcppEigen-package}}
}
